+++ buildSrc/src/main/groovy/ca/allanwang/kau/Plugins.groovy
@@ -9,4 +9,5 @@ class Plugins {
     static def androidMaven = "com.github.dcendents:android-maven-gradle-plugin:${Versions.mavenPlugin}"
     static def playPublisher = "com.github.triplet.gradle:play-publisher:${Versions.playPublishPlugin}"
     static def dexCount = "com.getkeepsafe.dexcount:dexcount-gradle-plugin:${Versions.dexCountPlugin}"
+    static def gitVersion = "gradle.plugin.com.gladed.gradle.androidgitversion:gradle-android-git-version:${Versions.gitVersionPlugin}"
 }