+++ buildSrc/src/main/groovy/ca/allanwang/kau/Versions.groovy
@@ -8,14 +8,14 @@ class Versions {
     static def buildTools = '27.0.2'
     static def supportLibs = '27.0.2'
 
-    static def kotlin = '1.2.20'
+    static def kotlin = '1.2.21'
 
     static def aboutLibraries = '6.0.1'
     static def anko = '0.10.4'
     static def blurry = '2.1.1'
     static def constraintLayout = '1.1.0-beta4'
-    static def fastAdapter = '3.1.2'
-    static def fastAdapterCommons = '3.1.2'
+    static def fastAdapter = '3.2.0'
+    static def fastAdapterCommons = '3.2.0'
     static def glide = '4.5.0'
 
     static def iconics = '3.0.2'
@@ -33,4 +33,5 @@ class Versions {
     static def mavenPlugin = '2.0'
     static def playPublishPlugin = '1.2.0'
     static def dexCountPlugin = '0.8.2'
+    static def gitVersionPlugin = '0.4.3'
 }