+++ docs/Changelog.md
@@ -5,6 +5,7 @@
 * :kpref-activity: Simplify internal code and add better encapsulation
 * Add Chinese, Indonesian, Norwegian, Polish, Thai, and Turkish translations
 * Add back git versioning
+* Created gradle plugin for getting version updates
 
 ## v3.6.2
 * :core: Pass null instead of bundle if bundle is empty for startActivity